fix bug in movie edit

+++ app/models/casting.rb
@@ -5,9 +5,11 @@ class Casting < ActiveRecord::Base
   belongs_to :movie
 
   def self.create_from_movie_form(actors, movie)
-    movie.castings.destroy_all
-    actors.each do |actor|
-      Actor.find_or_create(actor).castings.create(movie_id: movie.id)
+    unless movie.actors.map(&:name) == actors
+      movie.castings.destroy_all
+      actors.each do |actor|
+        Actor.find_or_create(actor).castings.create(movie_id: movie.id)
+      end
     end
   end
